refactor : observer 훅 분리
